% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinmarketcapr-package.r
\docType{package}
\name{coinmarketcapr}
\alias{coinmarketcapr}
\alias{coinmarketcapr-package}
\title{coinmarketcapr: Cryptocurrency Market Cap Prices from CoinMarketCap}
\description{
Extract and monitor price and market cap of 'Cryptocurrencies'
from 'CoinMarketCap' \url{https://coinmarketcap.com/api/} that lists many leading
cryptocurrencies along with their price, 24h trade volume, market cap and
much more in USD and other currencies.
}
\section{See Also}{

Useful links:
\itemize{
\item \url{https://coinmarketcap.com/api/}
\item \url{https://github.com/amrrs/coinmarketcapr}
\item Report bugs at \url{https://github.com/amrrs/coinmarketcapr/issues}
}
}

