% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coinmarketcapr.R
\name{get_marketcap_ticker_all}
\alias{get_marketcap_ticker_all}
\title{To extract Global Market Cap of Leading Cryptocurrencies}
\usage{
get_marketcap_ticker_all(currency = "USD")
}
\arguments{
\item{currency}{currency code - Default is 'USD'}
}
\value{
A dataframe of top Cryptocurrencies with Currency prices, Market Cap, Rank and more paramaters
}
\description{
To extract Global Market Cap of Leading Cryptocurrencies
}
\examples{
get_marketcap_ticker_all('EUR')
get_marketcap_ticker_all('GBP')
}

