% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\name{get_valid_currencies}
\alias{get_valid_currencies}
\title{Get Valid Currencies}
\usage{
get_valid_currencies()
}
\value{
A character vector of valid currencies supported by coinmarketcap API
}
\description{
Get Valid Currencies
}
\examples{
get_valid_currencies()
}
\seealso{
Other Cryptocurrencies: \code{\link{get_crypto_listings}},
  \code{\link{get_crypto_map}},
  \code{\link{get_crypto_marketpairs}},
  \code{\link{get_crypto_meta}},
  \code{\link{get_crypto_ohlcv}},
  \code{\link{get_crypto_quotes}},
  \code{\link{get_marketcap_ticker_all}}
}
\concept{Cryptocurrencies}
