% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinmarketcapr.R
\name{plot_top_currencies}
\alias{plot_top_currencies}
\title{Plot The Price of the Largest Market Cap Cryptocurrencies}
\usage{
plot_top_currencies(currency = "USD", k = 5)
}
\arguments{
\item{currency}{currency code - Default is 'USD'}

\item{k}{the number of top cryptocurrencies to plot (default is 5)}
}
\value{
A ggplot of top Cryptocurrencies based on their rank (Market Cap)
}
\description{
Plot The Price of the Largest Market Cap Cryptocurrencies
}
\examples{
plot_top_currencies('EUR')
plot_top_currencies('GBP')
}
