% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coinmarketcapr.R
\name{get_global_marketcap}
\alias{get_global_marketcap}
\title{To extract Global Market Cap of Leading Cryptocurrencies}
\usage{
get_global_marketcap(currency = "USD")
}
\arguments{
\item{currency}{currency code - Default is 'USD'}
}
\value{
A dataframe of Cryptocurrencies with Currency prices, Market Cap, Rank and more paramaters
}
\description{
To extract Global Market Cap of Leading Cryptocurrencies
}
\examples{
get_global_marketcap('AUD')
get_global_marketcap('EUR')
}

