% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinmarketcapr.R
\name{get_global_marketcap}
\alias{get_global_marketcap}
\title{Extract Global Market Cap of Cryptocurrency Market}
\usage{
get_global_marketcap(currency = "USD")
}
\arguments{
\item{currency}{currency code - Default is 'USD'}
}
\value{
A dataframe of to get global market cap of Cryptocurrencies Market
}
\description{
Extract Global Market Cap of Cryptocurrency Market
}
\examples{
get_global_marketcap('AUD')
get_global_marketcap('EUR')
}
