% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinmarketcapr.R
\name{get_marketcap_ticker_all}
\alias{get_marketcap_ticker_all}
\title{Extract Global Market Cap of Leading Cryptocurrencies}
\usage{
get_marketcap_ticker_all(currency = "USD")
}
\arguments{
\item{currency}{currency code - Default is 'USD'}
}
\value{
A dataframe of top Cryptocurrencies with id, name, symbol, rank, price_usd, price_btc, 24h_volume_usd, market_cap_usd, available_supply, total_supply, percent_change_1h, percent_change_24h, percent_change_7d, last_updated
}
\description{
Extract Global Market Cap of Leading Cryptocurrencies
}
\examples{
get_marketcap_ticker_all('EUR')
get_marketcap_ticker_all('GBP')
}
