% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\name{get_marketcap_ticker_all}
\alias{get_marketcap_ticker_all}
\title{Deprecated use \code{\link[coinmarketcapr]{get_crypto_listings}} instead.}
\usage{
get_marketcap_ticker_all(currency = "USD")
}
\arguments{
\item{currency}{currency code - Default is 'USD'}
}
\value{
A dataframe of top Cryptocurrencies with id, name, symbol, rank, price_usd, price_btc, 24h_volume_usd, market_cap_usd, available_supply, total_supply, percent_change_1h, percent_change_24h, percent_change_7d, last_updated
}
\description{
get_marketcap_ticker_all is replaced by \code{\link[coinmarketcapr]{get_crypto_listings}} to
support the new API and more functionality.
}
\examples{
\dontrun{
get_marketcap_ticker_all('EUR')
get_marketcap_ticker_all('GBP')
}
}
\seealso{
Other Cryptocurrencies: \code{\link{get_crypto_listings}},
  \code{\link{get_crypto_map}},
  \code{\link{get_crypto_marketpairs}},
  \code{\link{get_crypto_meta}},
  \code{\link{get_crypto_ohlcv}},
  \code{\link{get_crypto_quotes}},
  \code{\link{get_valid_currencies}}
}
\concept{Cryptocurrencies}
\keyword{internal}
