% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinmarketcapr.R
\name{get_valid_currencies}
\alias{get_valid_currencies}
\title{Get Valid Currencies}
\usage{
get_valid_currencies()
}
\value{
A character vector of valid currencies supported by coinmarketcap API
}
\description{
Get Valid Currencies
}
\examples{
get_valid_currencies()
}
