% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coinmarketcapr.R
\name{plot_top_5_currencies}
\alias{plot_top_5_currencies}
\title{To plot Top 5 Cryptocurrencies}
\usage{
plot_top_5_currencies(currency = "USD")
}
\arguments{
\item{currency}{currency code - Default is 'USD'}
}
\value{
A plot of top Cryptocurrencies based on their rank
}
\description{
To plot Top 5 Cryptocurrencies
}
\examples{
plot_top_5_currencies('EUR')
plot_top_5_currencies('GBP')
}

